refactor(renderer): extract ATTA attachment subsystem from renderer_objects
Move the building-attachment (ATTA) system into renderer_objects_atta.cpp: AttaOccupancyKey, SuppressAttachmentInMef, AddModelToLevelRes, UpdateAttaLocalPosInMef, LoadAttachmentsRecursive, DrawAttachmentsRecursive, DrawAttachmentsForSpline plus their file-local helpers (DisableAttaInMefBytes, AddEntriesToRes, ModelFamilyPrefix, StemInFamily). renderer_objects.cpp core drops 3646 -> 1918 lines. Pure code motion: builds clean, runtime-verified (level 2, 558 objects, 232 attachment ops, no errors).
