+++ CMakeLists.txt
@@ -77,6 +77,7 @@ set_source_files_properties(
   source/app.cpp
   source/renderer/renderer_flat_sky_layers.cpp
   source/renderer/renderer_objects.cpp
+  source/renderer/renderer_objects_atta.cpp
   source/renderer/renderer_objects_metadata.cpp
   source/renderer/renderer_objects_texture.cpp
   source/renderer/renderer_skydome.cpp
