+++ source/renderer/renderer_objects.cpp
@@ -517,559 +517,6 @@ void Renderer_Objects::InitPickingFBO(int w, int h) {
 // clicking anywhere on an ATTA part selects the owning LevelObject.
 // Mirrors DrawAttachmentsRecursive's transform calculation; skips transparency,
 // lighting, and texture setup since the picking shader only cares about geometry.
-std::string Renderer_Objects::AttaOccupancyKey(const std::string& modelId, const glm::vec3& worldPos) {
-    char buf[160];
-    snprintf(buf, sizeof(buf), "%s@%lld,%lld,%lld", modelId.c_str(),
-             (long long)llround(worldPos.x), (long long)llround(worldPos.y), (long long)llround(worldPos.z));
-    return buf;
-}
-
-// Disable a baked ATTA record inside an in-memory .mef byte buffer: zero its 16-byte
-// model NAME (so the engine can't resolve a model and skips it) and shove the record
-// far below the world as a fallback. Returns true if a matching record was edited.
-static bool DisableAttaInMefBytes(std::vector<uint8_t>& mef, const std::string& attModelId, const glm::vec3& localPos) {
-    if (mef.size() < 20 || std::memcmp(mef.data(), "ILFF", 4) != 0) return false;
-    bool modified = false;
-    size_t off = 20;
-    while (off + 16 <= mef.size()) {
-        char fourcc[5] = {0};
-        std::memcpy(fourcc, mef.data() + off, 4);
-        uint32_t size = 0, skip = 0;
-        std::memcpy(&size, mef.data() + off + 4, 4);
-        std::memcpy(&skip, mef.data() + off + 12, 4);
-
-        if (std::strcmp(fourcc, "ATTA") == 0) {
-            size_t recordsStart = off + 16;
-            size_t numRecords = size / 68;
-            for (size_t i = 0; i < numRecords; ++i) {
-                size_t base = recordsStart + i * 68;
-                if (base + 68 > mef.size()) break;
-                char nameBuf[16];
-                std::memcpy(nameBuf, mef.data() + base, 16);
-                float px, py, pz;
-                std::memcpy(&px, mef.data() + base + 16, 4);
-                std::memcpy(&py, mef.data() + base + 20, 4);
-                std::memcpy(&pz, mef.data() + base + 24, 4);
-                std::string aname(nameBuf, strnlen(nameBuf, 16));
-                if (aname == attModelId &&
-                    std::abs(px - localPos.x) < 1.0f &&
-                    std::abs(py - localPos.y) < 1.0f &&
-                    std::abs(pz - localPos.z) < 1.0f) {
-                    // Rename to a non-existent model ID so the game's resource
-                    // loader silently skips it (can't find the mesh, no render).
-                    // Must start with a digit — names starting with '_' or a letter
-                    // are treated as "magic object types" and show a warning dialog.
-                    // An empty name crashes IGI's loader entirely.
-                    static const char kNullName[16] = "999_99_9\0\0\0\0\0\0\0";
-                    std::memcpy(mef.data() + base, kNullName, 16);
-                    modified = true;
-                    break; // suppress only the first matching record per ATTA chunk
-                }
-            }
-        }
-        if (skip == 0) break;
-        off += skip;
-    }
-    return modified;
-}
-
-bool Renderer_Objects::SuppressAttachmentInMef(const std::string& parentModelId, const std::string& attModelId, const glm::vec3& localPos) {
-    // Patch the level's model archive IN PLACE: parse it, disable the one ATTA record
-    // inside the parent building's .mef entry, and repack ALL entries. This keeps the
-    // archive complete (the old code regenerated from editor/models/levelX, which the
-    // editor wipes after loading, producing an empty/broken .res).
-    const std::string gameRes = Utils::GetIGIRootPath() + "\\missions\\location0\\level" +
-                                std::to_string(current_level_) + "\\models\\level" +
-                                std::to_string(current_level_) + ".res";
-    if (!std::filesystem::exists(gameRes)) {
-        Logger::Get().Log(LogLevel::WARNING, "[Renderer] SuppressAttachmentInMef: archive not found: " + gameRes);
-        return false;
-    }
-
-    RESFile res = RES_Parse(gameRes);
-    if (!res.valid) {
-        Logger::Get().Log(LogLevel::ERR, "[Renderer] SuppressAttachmentInMef: failed to parse " + gameRes);
-        return false;
-    }
-
-    auto endsWithCI = [](const std::string& s, const std::string& suf) {
-        if (suf.size() > s.size()) return false;
-        for (size_t i = 0; i < suf.size(); ++i)
-            if (std::tolower((unsigned char)s[s.size()-suf.size()+i]) != std::tolower((unsigned char)suf[i])) return false;
-        return true;
-    };
-    const std::string suffix = parentModelId + ".mef";
-
-    bool modified = false;
-    for (auto& e : res.entries) {
-        if (endsWithCI(e.name, suffix)) {
-            if (DisableAttaInMefBytes(e.data, attModelId, localPos)) { modified = true; break; }
-        }
-    }
-    if (!modified) {
-        Logger::Get().Log(LogLevel::WARNING, "[Renderer] SuppressAttachmentInMef: ATTA '" + attModelId +
-            "' not found in " + parentModelId + ".mef");
-        return false;
-    }
-
-    try {
-        std::string bak = gameRes + ".orig";
-        if (!std::filesystem::exists(bak))
-            std::filesystem::copy_file(gameRes, bak, std::filesystem::copy_options::overwrite_existing);
-    } catch (...) {}
-
-    std::string err;
-    if (!RES_WriteEntries(res.entries, gameRes, err)) {
-        Logger::Get().Log(LogLevel::ERR, "[Renderer] SuppressAttachmentInMef: repack failed: " + err);
-        return false;
-    }
-    Logger::Get().Log(LogLevel::INFO, "[Renderer] SuppressAttachmentInMef: disabled ATTA '" + attModelId +
-        "' in " + parentModelId + ".mef and repacked " + gameRes);
-    return true;
-}
-
-// Batch-add resources (models or textures) to ONE .res archive, idempotently, WITHOUT
-// ever loading the whole archive into RAM. Streams the source once to discover which
-// requested names already exist (case-insensitive on the full entry name), then — if any
-// remain — streams it again appending only the missing entries into a .tmp and atomically
-// renames .tmp over the original. The original is backed up to <resPath>.orig once before
-// the first modification; if that backup fails we ABORT (archive untouched). On any
-// stream/IO/rename failure the .tmp is removed and the original is left intact.
-// This is the OOM-safe replacement for the old per-file RES_Parse → RES_WriteEntries path:
-// the 201MB textures archive is never held in memory (peak = one entry).
-static bool AddEntriesToRes(const std::string& resPath,
-                            const std::vector<RESEntry>& wanted,
-                            std::string& err,
-                            const std::function<void(size_t,size_t)>& onProgress = nullptr) {
-    auto equalsCI = [](const std::string& a, const std::string& b) {
-        if (a.size() != b.size()) return false;
-        for (size_t i = 0; i < a.size(); ++i)
-            if (std::tolower((unsigned char)a[i]) != std::tolower((unsigned char)b[i])) return false;
-        return true;
-    };
-
-    // Pass 1: stream source to find which wanted names are already present.
-    std::vector<bool> present(wanted.size(), false);
-    std::string ferr;
-    if (!RES_ForEachEntry(resPath,
-            [&](const std::string& name, const uint8_t*, size_t) {
-                for (size_t i = 0; i < wanted.size(); ++i)
-                    if (!present[i] && equalsCI(name, wanted[i].name)) present[i] = true;
-            }, ferr)) {
-        err = "membership scan failed: " + ferr;
-        return false;
-    }
-
-    std::vector<RESEntry> toAdd;
-    for (size_t i = 0; i < wanted.size(); ++i)
-        if (!present[i]) toAdd.push_back(wanted[i]);
-    if (toAdd.empty()) return true;  // everything already present -> no-op success
-
-    const std::string bak = resPath + ".orig";
-    if (!std::filesystem::exists(bak)) {
-        std::error_code ec;
-        std::filesystem::copy_file(resPath, bak, std::filesystem::copy_options::overwrite_existing, ec);
-        if (ec) {
-            err = "backup failed, aborting (archive untouched): " + ec.message();
-            return false;
-        }
-    }
-
-    const std::string tmp = resPath + ".tmp";
-    if (!RES_StreamAppend(resPath, toAdd, tmp, err, onProgress)) {
-        std::error_code rmec;
-        std::filesystem::remove(tmp, rmec);
-        return false;
-    }
-
-    std::error_code rec;
-    std::filesystem::rename(tmp, resPath, rec);
-    if (rec) {
-        err = "atomic replace failed (original untouched): " + rec.message();
-        std::error_code rmec;
-        std::filesystem::remove(tmp, rmec);
-        return false;
-    }
-    return true;
-}
-
-// Family prefix of a model id = the substring before the FIRST underscore
-// (e.g. "405_01_1" -> "405", "igiagent_1" -> "igiagent"). The whole MEF family
-// shares this prefix: hull + LODs + sub-parts are SEPARATE mefs the game needs
-// all of (e.g. 405_01_1, 405_01_2..5, 405_02_1.., 405_03_1..). If there is no
-// underscore the id is its own prefix.
-static std::string ModelFamilyPrefix(const std::string& modelId) {
-    const size_t us = modelId.find('_');
-    return (us == std::string::npos) ? modelId : modelId.substr(0, us);
-}
-
-// True if `stem` (a .mef filename without extension) belongs to family `prefix`,
-// i.e. it begins with exactly "<prefix>_". Scoped to the EXACT prefix-group so
-// "405" does not pull in "4050" or unrelated ids. Case-insensitive.
-static bool StemInFamily(const std::string& stem, const std::string& prefix) {
-    const std::string want = prefix + "_";
-    if (stem.size() <= want.size()) return false;
-    for (size_t i = 0; i < want.size(); ++i)
-        if (std::tolower((unsigned char)stem[i]) != std::tolower((unsigned char)want[i]))
-            return false;
-    return true;
-}
-
-bool Renderer_Objects::AddModelToLevelRes(const std::string& modelId,
-                                          const std::function<void(size_t,size_t)>& onProgress) {
-    const std::string levelDir = Utils::GetIGIRootPath() + "\\missions\\location0\\level" +
-        std::to_string(current_level_);
-    const std::string modelsRes   = levelDir + "\\models\\level"   + std::to_string(current_level_) + ".res";
-    const std::string texturesRes = levelDir + "\\textures\\level" + std::to_string(current_level_) + ".res";
-
-    if (!std::filesystem::exists(modelsRes)) {
-        Logger::Get().Log(LogLevel::WARNING, "[Renderer] AddModelToLevelRes: models archive missing: " + modelsRes);
-        return false;
-    }
-
-    // ── Enumerate the whole <prefix>_* MEF family on disk ────────────────────
-    // IGI models are families: adding only <modelId>.mef leaves the hull/LODs/
-    // sub-parts behind, which works on a level that already has them but breaks
-    // (transparent/crash) elsewhere. Scan the SAME dirs FindModelFile searches,
-    // collect every "<prefix>_*.mef", dedupe by base model id (stem).
-    const std::string prefix = ModelFamilyPrefix(modelId);
-
-    std::vector<std::string> familyDirs;
-    {
-        const std::string exeDir = Utils::GetExeDirectory();
-        const std::string lvl = std::to_string(current_level_);
-        familyDirs.push_back(exeDir + "\\editor\\models\\level" + lvl);
-        familyDirs.push_back(Utils::GetIGIModelsPath(current_level_));
-        for (int l = 1; l <= 14; ++l) {
-            if (l == current_level_) continue;
-            familyDirs.push_back(exeDir + "\\editor\\models\\level" + std::to_string(l));
-            familyDirs.push_back(Utils::GetIGIModelsPath(l));
-        }
-        familyDirs.push_back(exeDir + "\\editor\\models\\common");
-        familyDirs.push_back(Utils::GetIGIRootPath() + "\\missions\\location0\\common\\models");
-        familyDirs.push_back(Utils::GetIGIRootPath() + "\\editor\\models");
-    }
-
-    // Map base model id -> resolved .mef path (first dir wins; deduped by id).
-    std::map<std::string, std::string> familyModels; // ordered for stable logging
-    for (const std::string& dir : familyDirs) {
-        std::error_code ec;
-        if (!std::filesystem::exists(dir, ec)) continue;
-        for (const auto& entry : std::filesystem::directory_iterator(dir, ec)) {
-            if (ec) break;
-            if (!entry.is_regular_file()) continue;
-            const std::filesystem::path& p = entry.path();
-            if (p.extension() != ".mef") continue;
-            const std::string stem = p.stem().string();
-            if (!StemInFamily(stem, prefix)) continue;
-            familyModels.emplace(stem, p.string()); // keep first path seen for a given id
-        }
-    }
-
-    // ALWAYS include the originally-requested model even if the glob missed it.
-    if (!familyModels.count(modelId)) {
-        std::string mefPath = FindModelFile(modelId, /*isBuilding=*/false);
-        if (mefPath.empty()) mefPath = FindModelFile(modelId, true);
-        if (!mefPath.empty()) familyModels.emplace(modelId, mefPath);
-    }
-
-    if (familyModels.empty()) {
-        Logger::Get().Log(LogLevel::WARNING, "[Renderer] AddModelToLevelRes: no .mef found for family '" +
-            prefix + "' (requested " + modelId + ")");
-        return false;
-    }
-
-    // 1. Batch-add ALL family MEFs to the models archive (single streaming append).
-    std::vector<RESEntry> modelEntries;
-    std::vector<std::pair<std::string, std::vector<uint8_t>>> looseModels; // for editor-content copy
-    modelEntries.reserve(familyModels.size());
-    for (const auto& fm : familyModels) {
-        std::ifstream mf(fm.second, std::ios::binary);
-        std::vector<uint8_t> mefBytes((std::istreambuf_iterator<char>(mf)), std::istreambuf_iterator<char>());
-        if (mefBytes.empty()) {
-            Logger::Get().Log(LogLevel::INFO, "[Renderer] AddModelToLevelRes: empty/unreadable family .mef, skipping: " + fm.second);
-            continue;
-        }
-        looseModels.emplace_back(fm.first, mefBytes);
-        modelEntries.push_back(RESEntry{ "LOCAL:models/" + fm.first + ".mef", std::move(mefBytes) });
-    }
-    if (modelEntries.empty()) {
-        Logger::Get().Log(LogLevel::WARNING, "[Renderer] AddModelToLevelRes: all family .mef files empty/unreadable for '" + prefix + "'");
-        return false;
-    }
-
-    std::string err;
-    if (!AddEntriesToRes(modelsRes, modelEntries, err)) {
-        Logger::Get().Log(LogLevel::ERR, "[Renderer] AddModelToLevelRes: model add failed: " + err);
-        return false;
-    }
-    Logger::Get().Log(LogLevel::INFO, "[Renderer] AddModelToLevelRes: family '" + prefix + "' (" +
-        std::to_string(modelEntries.size()) + " mef) -> " + modelsRes);
-
-    // 2. Gather + batch-add ALL family textures into the textures archive in ONE
-    //    batched streaming append (the textures .res is 200MB+ — stream it once,
-    //    never per-texture, and never RES_Parse it into RAM).
-    const bool haveTexRes = std::filesystem::exists(texturesRes);
-    int texAdded = 0;
-    std::vector<std::pair<std::string, std::vector<uint8_t>>> looseTextures;  // for editor-content copy
-    std::vector<RESEntry> texEntries;
-    std::set<std::string> seenTex;
-    for (const auto& fm : familyModels) {
-        for (const std::string& texId : GetTextureIdsForModel(fm.first)) {
-            if (!seenTex.insert(texId).second) continue; // dedupe across the family
-            std::string texPath = FindTextureFile(texId);
-            if (texPath.empty()) {
-                Logger::Get().Log(LogLevel::INFO, "[Renderer] AddModelToLevelRes: texture " + texId +
-                    " not found on disk, skipping");
-                continue;
-            }
-            std::ifstream tf(texPath, std::ios::binary);
-            std::vector<uint8_t> texBytes((std::istreambuf_iterator<char>(tf)), std::istreambuf_iterator<char>());
-            if (texBytes.empty()) continue;
-            looseTextures.emplace_back(texId, texBytes);
-            texEntries.push_back(RESEntry{ "LOCAL:textures/" + texId + ".tex", std::move(texBytes) });
-        }
-    }
-
-    if (!texEntries.empty()) {
-        if (!haveTexRes) {
-            Logger::Get().Log(LogLevel::WARNING, "[Renderer] AddModelToLevelRes: textures archive missing, "
-                "skipping texture packing: " + texturesRes);
-            looseTextures.clear();
-        } else {
-            std::string terr;
-            if (AddEntriesToRes(texturesRes, texEntries, terr, onProgress)) {
-                texAdded = (int)texEntries.size();
-            } else {
-                Logger::Get().Log(LogLevel::WARNING, "[Renderer] AddModelToLevelRes: texture pack failed for family '" +
-                    prefix + "': " + terr);
-                looseTextures.clear();
-            }
-        }
-    }
-
-    // 3. Best-effort: copy loose files into the editor's content dirs so the editor
-    //    keeps resolving them after a level reload. Failure here is non-fatal.
-    try {
-        const std::string exeDir = Utils::GetExeDirectory();
-        const std::string lvl = std::to_string(current_level_);
-        std::filesystem::path modelDst = std::filesystem::path(exeDir) / "content" / "models" / ("level" + lvl);
-        std::filesystem::create_directories(modelDst);
-        for (const auto& m : looseModels) {
-            std::ofstream out((modelDst / (m.first + ".mef")).string(), std::ios::binary);
-            out.write(reinterpret_cast<const char*>(m.second.data()), (std::streamsize)m.second.size());
-        }
-        if (!looseTextures.empty()) {
-            std::filesystem::path texDst = std::filesystem::path(exeDir) / "content" / "textures" / ("level" + lvl);
-            std::filesystem::create_directories(texDst);
-            for (const auto& t : looseTextures) {
-                std::ofstream out((texDst / (t.first + ".tex")).string(), std::ios::binary);
-                out.write(reinterpret_cast<const char*>(t.second.data()), (std::streamsize)t.second.size());
-            }
-        }
-    } catch (const std::exception& e) {
-        Logger::Get().Log(LogLevel::WARNING, std::string("[Renderer] AddModelToLevelRes: editor-content copy failed: ") + e.what());
-    }
-
-    // 4. Register the model + its textures so the GAME can resolve the model's
-    //    materials (otherwise it renders transparent in-game). We update the TEXT
-    //    level<N>.dat then drive mtp_decoder.exe to regenerate the binary level<N>.mtp
-    //    (the game accepts the tool's output; a natively-written .mtp crashed it).
-    //    A failure here is a WARNING only -- the .res parts already succeeded.
-    {
-        const std::string lvl = std::to_string(current_level_);
-        const std::string lvlDir = Utils::GetIGIRootPath() +
-            "\\missions\\location0\\level" + lvl + "\\";
-        const std::string datPath = lvlDir + "level" + lvl + ".dat";
-        const std::string mtpPath = lvlDir + "level" + lvl + ".mtp";
-
-        if (!std::filesystem::exists(datPath)) {
-            Logger::Get().Log(LogLevel::WARNING, "[Renderer] AddModelToLevelRes: level .dat not found, "
-                "skipping model->texture mapping: " + datPath);
-        } else {
-            // 4a. Back up the .dat once, then add the model mapping and write it back.
-            bool proceed = true;
-            const std::string datBackup = datPath + ".orig";
-            if (!std::filesystem::exists(datBackup)) {
-                std::error_code ec;
-                std::filesystem::copy_file(datPath, datBackup,
-                    std::filesystem::copy_options::overwrite_existing, ec);
-                if (ec) {
-                    proceed = false;
-                    Logger::Get().Log(LogLevel::WARNING, "[Renderer] AddModelToLevelRes: could not back up "
-                        ".dat, skipping mapping update: " + datBackup + " (" + ec.message() + ")");
-                }
-            }
-
-            if (proceed) {
-                // Use the persistent in-memory DAT to avoid data loss caused by
-                // mtp_decoder.exe rewriting level.dat on disk between adds.
-                // On the first add this session, populate it from disk; on all
-                // subsequent adds for the same level, reuse the accumulated copy.
-                if (!persistent_dat_.valid || persistent_dat_path_ != datPath) {
-                    persistent_dat_ = DAT_Parse(datPath);
-                    persistent_dat_path_ = datPath;
-                    if (!persistent_dat_.valid) {
-                        Logger::Get().Log(LogLevel::WARNING, "[Renderer] AddModelToLevelRes: initial .dat parse failed: "
-                            + persistent_dat_.error + " — clearing persistent cache");
-                        persistent_dat_path_.clear();
-                    }
-                }
-                DATFile& dat = persistent_dat_;
-                if (!dat.valid) {
-                    Logger::Get().Log(LogLevel::WARNING, "[Renderer] AddModelToLevelRes: .dat parse failed, "
-                        "skipping mapping update: " + dat.error);
-                } else {
-                    // Add EVERY family model to the .dat; write once if any were new.
-                    bool anyAdded = false;
-                    for (const auto& fm : familyModels) {
-                        bool present = false;
-                        DAT_AddModel(dat, fm.first, GetTextureIdsForModel(fm.first), present);
-                        if (!present) anyAdded = true;
-                    }
-                    bool datReady = true; // either we'll write below, or it's already complete
-                    if (anyAdded) {
-                        std::string derr;
-                        if (DAT_WriteNative(dat, datPath, derr)) {
-                            Logger::Get().Log(LogLevel::INFO, "[Renderer] AddModelToLevelRes: added family '" +
-                                prefix + "' to " + datPath);
-                        } else {
-                            datReady = false;
-                            Logger::Get().Log(LogLevel::WARNING, "[Renderer] AddModelToLevelRes: .dat write "
-                                "failed for family '" + prefix + "': " + derr);
-                        }
-                    } else {
-                        Logger::Get().Log(LogLevel::INFO, "[Renderer] AddModelToLevelRes: family '" + prefix +
-                            "' already mapped in " + datPath);
-                    }
-
-                    // 4b. Back up the .mtp once, then add models natively via MTP_AddModel.
-                    if (datReady && std::filesystem::exists(mtpPath)) {
-                        const std::string mtpBackup = mtpPath + ".orig";
-                        if (!std::filesystem::exists(mtpBackup)) {
-                            std::error_code ec;
-                            std::filesystem::copy_file(mtpPath, mtpBackup,
-                                std::filesystem::copy_options::overwrite_existing, ec);
-                            if (ec)
-                                Logger::Get().Log(LogLevel::WARNING, "[Renderer] AddModelToLevelRes: "
-                                    "could not back up .mtp: " + mtpBackup + " (" + ec.message() + ")");
-                        }
-                        int mtpAdded = 0;
-                        for (const auto& fm : familyModels) {
-                            std::string merr;
-                            if (MTP_AddModel(mtpPath, mtpPath, fm.first, GetTextureIdsForModel(fm.first), merr)) {
-                                ++mtpAdded;
-                            } else {
-                                Logger::Get().Log(LogLevel::WARNING, "[Renderer] AddModelToLevelRes: "
-                                    "MTP_AddModel failed for " + fm.first + ": " + merr);
-                            }
-                        }
-                        Logger::Get().Log(LogLevel::INFO, "[Renderer] AddModelToLevelRes: updated " + mtpPath +
-                            " natively (" + std::to_string(mtpAdded) + " model(s) added/confirmed)");
-                    } else if (datReady) {
-                        Logger::Get().Log(LogLevel::WARNING, "[Renderer] AddModelToLevelRes: "
-                            ".mtp not found, skipping MTP update: " + mtpPath);
-                    }
-                }
-            }
-        }
-    }
-
-    Logger::Get().Log(LogLevel::INFO, "[Renderer] AddModelToLevelRes: packed family '" + prefix + "' = " +
-        std::to_string(modelEntries.size()) + " model(s) + " + std::to_string(texAdded) + " texture(s)");
-    return true;
-}
-
-bool Renderer_Objects::UpdateAttaLocalPosInMef(
-    const std::string& parentModelId, bool isBuilding,
-    int recordIndex, const glm::vec3& newLocalPos, const glm::mat3& newLocalRot)
-{
-    // 1. Update in-memory attachment cache so the editor renders at the new position.
-    const std::string prefix = isBuilding ? "building:" : "object:";
-    const std::string attKey = std::to_string(current_level_) + ":" + prefix + parentModelId;
-    auto cit = attachment_cache_.find(attKey);
-    if (cit != attachment_cache_.end() && recordIndex < (int)cit->second.size()) {
-        cit->second[recordIndex].px = newLocalPos.x;
-        cit->second[recordIndex].py = newLocalPos.y;
-        cit->second[recordIndex].pz = newLocalPos.z;
-    }
-
-    // 2. Patch px/py/pz bytes in the MEF entry inside the level .res.
-    const std::string gameRes = Utils::GetIGIRootPath() +
-        "\\missions\\location0\\level" + std::to_string(current_level_) +
-        "\\models\\level" + std::to_string(current_level_) + ".res";
-    if (!std::filesystem::exists(gameRes)) {
-        Logger::Get().Log(LogLevel::WARNING, "[Renderer] UpdateAttaLocalPosInMef: res not found: " + gameRes);
-        return false;
-    }
-    RESFile res = RES_Parse(gameRes);
-    if (!res.valid) return false;
-
-    auto endsWithCI = [](const std::string& s, const std::string& suf) {
-        if (suf.size() > s.size()) return false;
-        for (size_t i = 0; i < suf.size(); ++i)
-            if (std::tolower((unsigned char)s[s.size()-suf.size()+i]) !=
-                std::tolower((unsigned char)suf[i])) return false;
-        return true;
-    };
-
-    bool patched = false;
-    for (auto& e : res.entries) {
-        if (!endsWithCI(e.name, parentModelId + ".mef")) continue;
-        size_t off = 20; // skip ILFF(16) + IRES(4)
-        while (off + 16 <= e.data.size()) {
-            char fc[5] = {0}; std::memcpy(fc, e.data.data() + off, 4);
-            uint32_t size = 0, skip = 0;
-            std::memcpy(&size, e.data.data() + off + 4, 4);
-            std::memcpy(&skip, e.data.data() + off + 12, 4);
-            if (std::strcmp(fc, "ATTA") == 0) {
-                size_t base = off + 16 + (size_t)recordIndex * 68;
-                if (base + 68 <= e.data.size()) {
-                    // Write position
-                    std::memcpy(e.data.data() + base + 16, &newLocalPos.x, 4);
-                    std::memcpy(e.data.data() + base + 20, &newLocalPos.y, 4);
-                    std::memcpy(e.data.data() + base + 24, &newLocalPos.z, 4);
-                    // Write rotation matrix: MEF stores r[9] as column-major floats
-                    // matching the GLM mat3 column layout [col][row].
-                    float rot9[9] = {
-                        newLocalRot[0][0], newLocalRot[0][1], newLocalRot[0][2],
-                        newLocalRot[1][0], newLocalRot[1][1], newLocalRot[1][2],
-                        newLocalRot[2][0], newLocalRot[2][1], newLocalRot[2][2]
-                    };
-                    std::memcpy(e.data.data() + base + 28, rot9, 36);
-                    patched = true;
-                }
-                break;
-            }
-            if (skip == 0) break;
-            off += skip;
-        }
-        if (patched) break;
-    }
-    if (!patched) {
-        Logger::Get().Log(LogLevel::WARNING, "[Renderer] UpdateAttaLocalPosInMef: record " +
-            std::to_string(recordIndex) + " not found in " + parentModelId + ".mef");
-        return false;
-    }
-
-    try {
-        std::string bak = gameRes + ".orig";
-        if (!std::filesystem::exists(bak))
-            std::filesystem::copy_file(gameRes, bak, std::filesystem::copy_options::overwrite_existing);
-    } catch (...) {}
-
-    std::string err;
-    if (!RES_WriteEntries(res.entries, gameRes, err)) {
-        Logger::Get().Log(LogLevel::ERR, "[Renderer] UpdateAttaLocalPosInMef: repack failed: " + err);
-        return false;
-    }
-    Logger::Get().Log(LogLevel::INFO, "[Renderer] UpdateAttaLocalPosInMef: patched record " +
-        std::to_string(recordIndex) + " in " + parentModelId + ".mef -> local(" +
-        std::to_string(newLocalPos.x) + "," + std::to_string(newLocalPos.y) + "," +
-        std::to_string(newLocalPos.z) + ")");
-    return true;
-}
-
 void Renderer_Objects::DrawAttachmentsForPicking(
     const std::string& parentModelId, bool isBuilding,
     const glm::mat4& parentWorldMat, float parentScale,
@@ -1881,319 +1328,6 @@ void Renderer_Objects::DrawModelPreview(const std::string& modelId, GLuint ubo_m
 // Recursively scans the ATTA section of modelId's MEF file, caches each
 // sub-model mesh AND its own attachment records, then recurses into children.
 // The visited set prevents infinite loops on circular ATTA references.
-void Renderer_Objects::LoadAttachmentsRecursive(const std::string& modelId, bool isBuilding,
-                                                 std::unordered_set<std::string>& visited) {
-    if (!visited.insert(modelId).second) return; // cycle guard
-
-    std::string cacheKey = std::to_string(current_level_) + ":" +
-                           (isBuilding ? "building:" : "object:") + modelId;
-
-    if (attachment_cache_.find(cacheKey) != attachment_cache_.end()) return;
-
-    std::string filepath = FindModelFile(modelId, isBuilding);
-    if (filepath.empty()) {
-        attachment_cache_[cacheKey] = {};
-        return;
-    }
-
-    std::vector<AttachInfo> attaches;
-    try {
-        ParsedGeometry geo = ParseMefFile(filepath);
-        for (const auto& a : geo.mefAttachments) {
-            std::string aname(a.name, strnlen(a.name, 16));
-            if (aname.empty()) continue;
-
-            AttachInfo info;
-            info.modelId = aname;
-            info.px = a.px; info.py = a.py; info.pz = a.pz;
-            info.r[0]=a.r00; info.r[1]=a.r01; info.r[2]=a.r02;
-            info.r[3]=a.r03; info.r[4]=a.r04; info.r[5]=a.r05;
-            info.r[6]=a.r06; info.r[7]=a.r07; info.r[8]=a.r08;
-            attaches.push_back(info);
-
-            Logger::Get().Log(LogLevel::INFO,
-                "[Renderer_Objects] Attachment '" + aname + "' of '" + modelId +
-                "' pos=(" + std::to_string(a.px) + "," + std::to_string(a.py) +
-                "," + std::to_string(a.pz) + ")");
-
-            // Pre-warm the sub-model mesh
-            std::string subFile = FindModelFile(aname, isBuilding);
-            if (subFile.empty()) {
-                Logger::Get().Log(LogLevel::WARNING,
-                    "[Renderer_Objects] Attachment sub-model NOT FOUND: " + aname);
-            } else {
-                std::string subKey = std::to_string(current_level_) + ":" +
-                                     (isBuilding ? "building:" : "object:") + aname;
-                if (mesh_cache_.find(subKey) == mesh_cache_.end()) {
-                    try {
-                        Mesh subMesh = loadObjModel(subFile, "");
-                        ApplyTexturesToMesh(subMesh, aname, modelId);
-                        mesh_cache_[subKey] = subMesh;
-                        Logger::Get().Log(LogLevel::INFO,
-                            "[Renderer_Objects] Attachment sub-model loaded: " + aname +
-                            " (" + std::to_string(subMesh.vertexCount) + " verts)");
-                    } catch (const std::exception &se) {
-                        Logger::Get().Log(LogLevel::ERR,
-                            "[Renderer_Objects] Attachment sub-model load FAILED: " + aname + ": " + se.what());
-                        Mesh empty; mesh_cache_[subKey] = empty;
-                    }
-                }
-                // Recurse: parse this child's own ATTA section
-                LoadAttachmentsRecursive(aname, isBuilding, visited);
-            }
-        }
-    } catch (const std::exception &pe) {
-        Logger::Get().Log(LogLevel::WARNING,
-            "[Renderer_Objects] Could not parse ATTA from '" + filepath + "': " + pe.what());
-    }
-
-    attachment_cache_[cacheKey] = std::move(attaches);
-    Logger::Get().Log(LogLevel::INFO,
-        "[Renderer_Objects] Attachments for '" + modelId + "': " +
-        std::to_string(attachment_cache_[cacheKey].size()));
-}
-
-// ─── DrawAttachmentsRecursive ────────────────────────────────────────────────
-// Draws all ATTA children of parentModelId, then recurses into each child's
-// own attachments. parentWorldMat is the UN-SCALED world transform of the parent
-// (translate + rotate only), so children can position themselves relative to it.
-// The 40.96 scale is applied only at the leaf draw call.
-void Renderer_Objects::DrawAttachmentsRecursive(
-    const std::string& topLevelModelId, const std::string& parentModelId, bool isBuilding, const glm::mat4& parentWorldMat,
-    bool isTransparentPass, GLint loc_model, GLint loc_dirlight,
-    GLint loc_ambient, GLint loc_useTex, GLint loc_tex, GLint loc_alpha,
-    std::unordered_set<std::string>& drawn,
-    glm::vec3 leafScale)
-{
-    // Skip rendering attachments for any weapon model
-    if (IsWeaponModel(parentModelId)) return;
-    std::string prefix = isBuilding ? "building:" : "object:";
-    std::string attCacheKey = std::to_string(current_level_) + ":" + prefix + parentModelId;
-    auto ait = attachment_cache_.find(attCacheKey);
-    if (ait == attachment_cache_.end()) return;
-
-    const auto& attsR = ait->second;
-    for (size_t rri = 0; rri < attsR.size(); ++rri) {
-        const auto &att = attsR[rri];
-        // Find the mesh
-        std::string subKey = std::to_string(current_level_) + ":" + prefix + att.modelId;
-        auto sit = mesh_cache_.find(subKey);
-        if (sit == mesh_cache_.end() || sit->second.vertexCount == 0) {
-            subKey = std::to_string(current_level_) + ":object:" + att.modelId;
-            sit = mesh_cache_.find(subKey);
-        }
-        if (sit == mesh_cache_.end() || sit->second.vertexCount == 0) continue;
-        const Mesh &subMesh = sit->second;
-
-        // Skip ATTA sub-models that are TASKTYPE_DEATHZONE magic objects — they are
-        // invisible trigger/boarding zones, not visual geometry.
-        EnsureDeathZoneIdsLoaded();
-        if (deathzone_ids_.count(att.modelId)) continue;
-
-        // Build the ATTA local rotation matrix (DirectX row-major → GLM column-major)
-        glm::mat4 attLocalRot(
-            att.r[0], att.r[1], att.r[2], 0.f,
-            att.r[3], att.r[4], att.r[5], 0.f,
-            att.r[6], att.r[7], att.r[8], 0.f,
-            0.f,      0.f,      0.f,      1.f
-        );
-
-        // ATTA px/py/pz are raw floats from the MEF file, in the same coordinate
-        // space as obj.pos (raw game units). parentWorldMat is also in raw game units
-        // (translate=obj.pos, no scale), so no conversion needed.
-        glm::vec3 localOff(att.px, att.py, att.pz);
-        glm::vec3 worldPos = glm::vec3(parentWorldMat * glm::vec4(localOff, 1.f));
-
-        // Extract parent rotation (upper-left 3x3 of the unscaled parent mat)
-        glm::mat4 parentRot = parentWorldMat;
-        parentRot[3] = glm::vec4(0.f, 0.f, 0.f, 1.f); // zero out translation
-
-        // Unscaled world matrix for this attachment (used by children)
-        glm::mat4 childWorldMat(1.0f);
-        childWorldMat = glm::translate(childWorldMat, worldPos);
-        childWorldMat = childWorldMat * parentRot * attLocalRot;
-
-        // If this ATTA has a proxy object editing it, skip rendering it here —
-        // the proxy renders it — but still recurse into children.
-        if (IsAttaPromoted(att.modelId, worldPos) ||
-            promoted_atta_records_.count(parentModelId + ":" + std::to_string(rri)) > 0) {
-            std::string childKey = parentModelId + ">" + att.modelId;
-            if (drawn.insert(childKey).second) {
-                DrawAttachmentsRecursive(topLevelModelId, att.modelId, isBuilding, childWorldMat, isTransparentPass,
-                                         loc_model, loc_dirlight, loc_ambient,
-                                         loc_useTex, loc_tex, loc_alpha, drawn, leafScale);
-            }
-            continue;
-        }
-
-        // Skip sub-models that have only collision fallback geometry (no XTRV/DNER render
-        // vertices). These render as misshapen boxes with fabricated UVs.  Still recurse
-        // into their children — those may have proper render geometry.
-        if (!subMesh.fromRenderMesh) {
-            std::string childKey = parentModelId + ">" + att.modelId;
-            if (drawn.insert(childKey).second) {
-                DrawAttachmentsRecursive(topLevelModelId, att.modelId, isBuilding, childWorldMat, isTransparentPass,
-                                         loc_model, loc_dirlight, loc_ambient,
-                                         loc_useTex, loc_tex, loc_alpha, drawn, leafScale);
-            }
-            continue;
-        }
-
-        // Skip untextured sub-models (trigger zones, boarding areas, collision triggers)
-        // — they'd appear as featureless dark/gray slabs with no visual value.
-        // Exception: small untextured sub-models (e.g. sunglasses and other character
-        // head/body accessories) are legitimate visual geometry — they carry no DAT texture
-        // entry but should still draw via the hash-color fallback path below. Trigger/boarding
-        // slabs span several metres (half-extent in the ~1000s of raw units); accessory
-        // geometry is head-sized (a few hundred units at most), so a max-half-extent threshold
-        // separates them cleanly without re-introducing the slabs.
-        {
-            bool subHasTex = (subMesh.textureID > 0);
-            for (const auto& s : subMesh.subMeshes) {
-                if (s.textureID > 0) { subHasTex = true; break; }
-            }
-            const float kSlabHalfExtent = 700.0f; // ~0.17 m: above any accessory, below any slab
-            const float subMaxHalfExtent = std::max(subMesh.halfExtents.x,
-                std::max(subMesh.halfExtents.y, subMesh.halfExtents.z));
-            const bool isLargeSlab = (subMaxHalfExtent >= kSlabHalfExtent);
-            EnsureAiModelIdsLoaded();
-            const bool isAiModel = (ai_model_ids_.count(topLevelModelId) > 0);
-            if (!subHasTex && (!isAiModel || isLargeSlab)) {
-                std::string childKey = parentModelId + ">" + att.modelId;
-                if (drawn.insert(childKey).second) {
-                    DrawAttachmentsRecursive(topLevelModelId, att.modelId, isBuilding, childWorldMat, isTransparentPass,
-                                             loc_model, loc_dirlight, loc_ambient,
-                                             loc_useTex, loc_tex, loc_alpha, drawn, leafScale);
-                }
-                continue;
-            }
-        }
-
-        // Scaled model matrix: X uses leafScale.x (intervalLen for splines, 40.96 for buildings)
-        glm::mat4 attModel = glm::scale(childWorldMat, leafScale);
-
-        glUniformMatrix4fv(loc_model, 1, GL_FALSE, glm::value_ptr(attModel));
-
-        // Determine rendering pass requirements:
-        //   attIsWindow: named "WINDOW"/"GLASS" → transparent pass only
-        //   attHasAlpha: any submesh has alphaMode==2 (DNER rawOpacity non-zero)
-        //                → draw in BOTH passes (opaque pass discards α<0.5 pixels,
-        //                  transparent pass blends the rest)
-        const bool attIsWindow = window_model_ids_.count(att.modelId) > 0;
-        const bool attIsTree = (att.modelId == "900_01_1" ||
-                                att.modelId == "902_01_1" ||
-                                att.modelId == "905_01_1");
-        bool attHasAlpha = false;
-        if (!attIsTree) {
-            for (const auto& s : subMesh.subMeshes) {
-                if (s.alphaMode == 2) { attHasAlpha = true; break; }
-            }
-        }
-        if (current_level_ == 12 && !attIsWindow) {
-            attHasAlpha = false;
-        }
-
-        // Skip this sub-model if it doesn't belong in the current pass:
-        //   Windows: transparent pass only
-        //   Alpha attachments: both passes
-        //   Opaque attachments: opaque pass only
-        const bool drawInThisPass = attIsWindow
-            ? isTransparentPass
-            : (!isTransparentPass || attHasAlpha);
-        if (!drawInThisPass) {
-            // Still recurse into children — they may need a different pass
-            std::string childKey = parentModelId + ">" + att.modelId;
-            if (drawn.insert(childKey).second) {
-                DrawAttachmentsRecursive(topLevelModelId, att.modelId, isBuilding, childWorldMat, isTransparentPass,
-                                          loc_model, loc_dirlight, loc_ambient,
-                                          loc_useTex, loc_tex, loc_alpha, drawn, leafScale);
-            }
-            continue;
-        }
-
-        if (attIsWindow) {
-            glEnable(GL_BLEND);
-            glBlendFunc(GL_SRC_ALPHA, GL_ONE_MINUS_SRC_ALPHA);
-            glUniform1f(loc_alpha, 0.55f); // visible reflectivity, still see-through
-            glUniform1f(loc_glass_min_, 0.30f); // clean glass sheen floor
-        } else {
-            glUniform1f(loc_alpha, 1.0f);
-            glUniform1f(loc_glass_min_, 0.0f);
-        }
-
-        // Draw sub-model submeshes
-        if (!subMesh.subMeshes.empty()) {
-            for (const auto &sub : subMesh.subMeshes) {
-                if (sub.VAO == 0 || sub.vertexCount == 0) continue;
-
-                // In the transparent pass, only draw alpha submeshes (with blending).
-                // In the opaque pass, draw everything — the shader discards α<0.5 pixels.
-                bool subNeedsBlend = (sub.alphaMode == 2);
-                if (current_level_ == 12 && !attIsWindow) {
-                    subNeedsBlend = false;
-                }
-                if (attIsTree) subNeedsBlend = false;
-                if (isTransparentPass && !attIsWindow && !subNeedsBlend) continue;
-
-                if (subNeedsBlend) {
-                    glEnable(GL_BLEND);
-                    glBlendFunc(GL_SRC_ALPHA, GL_ONE_MINUS_SRC_ALPHA);
-                    glUniform1f(loc_alpha, sub.baseColorFactor.a);
-                }
-
-                if (sub.textureID > 0) {
-                    // Same normal lighting for window/glass attachments so they stay
-                    // clear and see-through (see top-level path — flat-gray reverted).
-                    glUniform3f(loc_dirlight, 0.6f, 0.6f, 0.6f);
-                    glUniform3f(loc_ambient,  0.4f, 0.4f, 0.4f);
-                    glUniform1i(loc_useTex, 1);
-                    glActiveTexture(GL_TEXTURE0);
-                    glBindTexture(GL_TEXTURE_2D, sub.textureID);
-                    glUniform1i(loc_tex, 0);
-                } else {
-                    // Untextured accessory (e.g. sunglasses): hash-color fallback, matching
-                    // the untextured submesh path used for top-level objects above.
-                    size_t hash = std::hash<std::string>{}(att.modelId);
-                    float hr = 0.4f + (float)(hash & 0xFF) / 255.0f * 0.4f;
-                    float hg = 0.4f + (float)((hash >> 8) & 0xFF) / 255.0f * 0.4f;
-                    float hb = 0.4f + (float)((hash >> 16) & 0xFF) / 255.0f * 0.4f;
-                    glUniform3f(loc_dirlight, hr * 0.6f, hg * 0.6f, hb * 0.6f);
-                    glUniform3f(loc_ambient,  hr * 0.4f, hg * 0.4f, hb * 0.4f);
-                    glUniform1i(loc_useTex, 0);
-                }
-                glBindVertexArray(sub.VAO);
-                glDrawArrays(GL_TRIANGLES, 0, sub.vertexCount);
-
-                if (subNeedsBlend) {
-                    glDisable(GL_BLEND);
-                    glUniform1f(loc_alpha, 1.0f);
-                }
-            }
-            glBindVertexArray(0);
-        } else if (subMesh.textureID > 0) {
-            glUniform3f(loc_dirlight, 0.6f, 0.6f, 0.6f);
-            glUniform3f(loc_ambient,  0.4f, 0.4f, 0.4f);
-            glUniform1i(loc_useTex, 1);
-            GL_BindTexture2D(0, subMesh.textureID);
-            glUniform1i(loc_tex, 0);
-            renderModel(subMesh);
-        }
-
-        if (attIsWindow) {
-            glDisable(GL_BLEND);
-            glUniform1f(loc_alpha, 1.0f);
-        }
-
-        // Recurse into this attachment's own children
-        std::string childKey = parentModelId + ">" + att.modelId;
-        if (drawn.insert(childKey).second) {
-            DrawAttachmentsRecursive(topLevelModelId, att.modelId, isBuilding, childWorldMat, isTransparentPass,
-                                      loc_model, loc_dirlight, loc_ambient,
-                                      loc_useTex, loc_tex, loc_alpha, drawn, leafScale);
-        }
-    }
-}
-
 bool Renderer_Objects::IsVehicleType(const std::string& type) {
     return type == "Car" || type == "Heli" || type == "Plane" || type == "Train";
 }
@@ -2782,51 +1916,3 @@ void main() {
 // unscaledWorldMat = translate + rotate at the segment midpoint (no scale).
 // leafScale.x = intervalLen (stretch ATTA to cover the full segment in X).
 // leafScale.y/z = 40.96 (normal world scale for width/height).
-void Renderer_Objects::DrawAttachmentsForSpline(
-    const std::string& modelId, bool isBuilding,
-    const glm::mat4& unscaledWorldMat, GLuint ubo_mats,
-    glm::vec3 leafScale)
-{
-    EnsureWindowModelIdsLoaded();
-
-    glUseProgram(shader_program_);
-    glBindBufferBase(GL_UNIFORM_BUFFER, ubo_binding_point_, ubo_mats);
-
-    GLint loc_model    = glGetUniformLocation(shader_program_, "u_model");
-    GLint loc_dirlight = glGetUniformLocation(shader_program_, "u_dirlight");
-    GLint loc_ambient  = glGetUniformLocation(shader_program_, "u_ambient");
-    GLint loc_useTex   = glGetUniformLocation(shader_program_, "u_useTexture");
-    GLint loc_tex      = glGetUniformLocation(shader_program_, "u_texture");
-    GLint loc_alpha    = glGetUniformLocation(shader_program_, "u_alpha");
-
-    glUniform1f(loc_alpha, 1.0f);
-
-    glEnable(GL_POLYGON_OFFSET_FILL);
-    glPolygonOffset(-2.0f, -2.0f);
-
-    // Opaque pass
-    {
-        std::unordered_set<std::string> drawn;
-        DrawAttachmentsRecursive(modelId, modelId, isBuilding, unscaledWorldMat, /*isTransparentPass=*/false,
-                                 loc_model, loc_dirlight, loc_ambient, loc_useTex, loc_tex, loc_alpha, drawn, leafScale);
-    }
-
-    // Transparent pass (windows / glass)
-    {
-        std::unordered_set<std::string> drawn;
-        DrawAttachmentsRecursive(modelId, modelId, isBuilding, unscaledWorldMat, /*isTransparentPass=*/true,
-                                 loc_model, loc_dirlight, loc_ambient, loc_useTex, loc_tex, loc_alpha, drawn, leafScale);
-    }
-
-    // Restore blend/depth state — transparent pass may leave them dirty.
-    glDisable(GL_BLEND);
-    glDepthMask(GL_TRUE);
-    glUniform1f(loc_alpha, 1.0f);
-    glDisable(GL_POLYGON_OFFSET_FILL);
-    glUseProgram(0);
-}
-
-
-
-
-
